% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-logspline.R
\docType{data}
\name{StatLogspline}
\alias{StatLogspline}
\title{Base ggproto classes for gglogspline}
\format{An object of class \code{StatLogspline} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
StatLogspline
}
\description{
Base ggproto classes for gglogspline
}
\section{Computed variables}{

\itemize{
\item \code{density} : the density estimate
\item \code{count}: computed counts (similar to \code{\link[ggplot2:stat_density]{ggplot2::stat_density()}})
\item \code{probs}: distribution function
\item \code{survival}: survival function
\item \code{hazard} : hazard function
}

By default the \code{y} aesthetic is mapped to \code{stat(density)}
}

\keyword{datasets}
