% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglogspline-package.R
\docType{package}
\name{gglogspline}
\alias{gglogspline}
\alias{gglogspline-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/gglogspline}
\item BugReports: \url{https://gitlab.com/hrbrmstr/gglogspline/issues}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/gglogspline}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/gglogspline/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
