% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-logspline.R
\name{stat_logspline}
\alias{stat_logspline}
\title{Computes logspline density (+ counts estimate), probability, survival & hazard}
\usage{
stat_logspline(mapping = NULL, data = NULL, geom = "area",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, n = 100, max_knots = 0, n_knots = 0,
  min_d = -1, error_action = 2, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{Use to override the default connection between
\code{geom_density} and \code{stat_density}.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{n}{numbe of points for the density estimation (larger == smoother)}

\item{max_knots}{the maximum number of knots. The routine stops adding knots when
this number of knots is reached. The method has an automatic rule for selecting
maxknots if this parameter is not specified.}

\item{n_knots}{forces the method to start with nknots knots. The method has an automatic
rule for selecting nknots if this parameter is not specified.}

\item{min_d}{minimum distance, in order statistics, between knots.}

\item{error_action}{see \code{error.action} in \code{\link[logspline:plot.logspline]{logspline::plot.logspline()}}}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Computes logspline density (+ counts estimate), probability, survival & hazard
}
\section{Computed variables}{

\itemize{
\item \code{density} : the density estimate
\item \code{count}: computed counts (similar to \code{\link[ggplot2:stat_density]{ggplot2::stat_density()}})
\item \code{probs}: distribution function
\item \code{survival}: survival function
\item \code{hazard} : hazard function
}

By default the \code{y} aesthetic is mapped to \code{stat(density)}
}

\examples{
library(ggplot2)

set.seed(1)
data.frame(
  val = rnorm(100)
) -> xdf

ggplot(xdf) + stat_logspline(aes(val))
}
