% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglogspline-package.R
\docType{package}
\name{gglogspline}
\alias{gglogspline}
\alias{gglogspline-package}
\title{A 'ggplot2' Extension for Visualizing Density, Distribution, Hazard, or Survival
Functions using the 'logspline' Package}
\description{
Methods are provided to plot a logspline density, distribution function,
hazard function or survival function from a logspline density
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/gglogspline}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/gglogspline/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
